Renamed contrib to sockeye_contrib
