+++ MANIFEST.in
@@ -10,8 +10,8 @@ include sockeye/git_version.py
 recursive-include .github *
 include CONTRIBUTING.md
 exclude *.sh
-exclude contrib
-recursive-exclude contrib *
+exclude sockeye_contrib
+recursive-exclude sockeye_contrib *
 include pytest.ini
 recursive-include test *.py
 recursive-include test *.ini
