+++ sockeye_contrib/autopilot/autopilot.py
@@ -37,15 +37,15 @@ except ImportError:
     raise RuntimeError("Please install the sockeye module or add the sockeye root directory to your Python path. Ex: export PYTHONPATH=%s"
                        % SOCKEYE_ROOT)
 
-from contrib.autopilot.tasks import ARCHIVE_NONE, ARCHIVE_TAR, ARCHIVE_ZIP
-from contrib.autopilot.tasks import TEXT_UTF8_RAW, TEXT_UTF8_RAW_SGML, TEXT_UTF8_RAW_BITEXT
-from contrib.autopilot.tasks import TEXT_UTF8_RAW_BITEXT_REVERSE, TEXT_REQUIRES_TOKENIZATION
-from contrib.autopilot.tasks import TEXT_UTF8_TOKENIZED
-from contrib.autopilot.tasks import RAW_FILES
-from contrib.autopilot.tasks import Task, TASKS
-from contrib.autopilot.models import MODELS, MODEL_NONE, MODEL_GNMT, MODEL_TEST_ARGS
-from contrib.autopilot.models import DECODE_ARGS, DECODE_STANDARD, DECODE_GNMT
-from contrib.autopilot import third_party
+from sockeye_contrib.autopilot.tasks import ARCHIVE_NONE, ARCHIVE_TAR, ARCHIVE_ZIP
+from sockeye_contrib.autopilot.tasks import TEXT_UTF8_RAW, TEXT_UTF8_RAW_SGML, TEXT_UTF8_RAW_BITEXT
+from sockeye_contrib.autopilot.tasks import TEXT_UTF8_RAW_BITEXT_REVERSE, TEXT_REQUIRES_TOKENIZATION
+from sockeye_contrib.autopilot.tasks import TEXT_UTF8_TOKENIZED
+from sockeye_contrib.autopilot.tasks import RAW_FILES
+from sockeye_contrib.autopilot.tasks import Task, TASKS
+from sockeye_contrib.autopilot.models import MODELS, MODEL_NONE, MODEL_GNMT, MODEL_TEST_ARGS
+from sockeye_contrib.autopilot.models import DECODE_ARGS, DECODE_STANDARD, DECODE_GNMT
+from sockeye_contrib.autopilot import third_party
 
 
 # Formats for custom files
@@ -457,7 +457,7 @@ def call_sockeye_translate(args: List[str],
 
 def call_sacrebleu(input_fname: str, ref_fname: str, output_fname: str, log_fname: str, tokenized: bool = False):
     """
-    Call contrib.sacrebleu.sacrebleu on tokenized or detokenized inputs.
+    Call sockeye_contrib.sacrebleu.sacrebleu on tokenized or detokenized inputs.
 
     :param input_fname: Input translation file.
     :param ref_fname: Reference translation file.
@@ -468,7 +468,7 @@ def call_sacrebleu(input_fname: str, ref_fname: str, output_fname: str, log_fnam
     # Assemble command
     command = [sys.executable,
                "-m",
-               "contrib.sacrebleu.sacrebleu",
+               "sockye_contrib.sacrebleu.sacrebleu",
                "--score-only",
                "--input={}".format(input_fname),
                ref_fname]
@@ -477,7 +477,7 @@ def call_sacrebleu(input_fname: str, ref_fname: str, output_fname: str, log_fnam
         command.append("--tokenize=none")
     # Call sacrebleu
     with open(log_fname, "wb") as log:
-        logging.info("contrib.sacrebleu.sacrebleu: %s -> %s", input_fname, output_fname)
+        logging.info("sockeye_contrib.sacrebleu.sacrebleu: %s -> %s", input_fname, output_fname)
         logging.info("Log: %s", log_fname)
         score = subprocess.check_output(command, stderr=log)
     # Record successful score
@@ -796,7 +796,7 @@ def run_steps(args: argparse.Namespace):
         else:
             fname_log = os.path.join(args.workspace,
                          DIR_LOGS,
-                         "contrib.sacrebleu.sacrebleu.{}.{}.{}.{}.log".format(task_name,
+                         "sockeye_contrib.sacrebleu.sacrebleu.{}.{}.{}.{}.log".format(task_name,
                                                                               args.model,
                                                                               fname_base + SUFFIX_BPE,
                                                                               os.getpid()))
@@ -817,7 +817,7 @@ def run_steps(args: argparse.Namespace):
             third_party.merge_bpe(input_fname=fname_bpe, output_fname=fname_tok)
             fname_log = os.path.join(args.workspace,
                          DIR_LOGS,
-                         "contrib.sacrebleu.sacrebleu.{}.{}.{}.{}.log".format(task_name,
+                         "sockeye_contrib.sacrebleu.sacrebleu.{}.{}.{}.{}.log".format(task_name,
                                                                               args.model,
                                                                               fname_base + SUFFIX_TOK,
                                                                               os.getpid()))
@@ -844,7 +844,7 @@ def run_steps(args: argparse.Namespace):
                                                lang_code=lang_code)
             fname_log = os.path.join(args.workspace,
                          DIR_LOGS,
-                         "contrib.sacrebleu.sacrebleu.{}.{}.{}.{}.log".format(task_name,
+                         "sockeye_contrib.sacrebleu.sacrebleu.{}.{}.{}.{}.log".format(task_name,
                                                                               args.model,
                                                                               fname_base + SUFFIX_DETOK,
                                                                               os.getpid()))
