+++ sockeye_contrib/autopilot/test.py
@@ -23,7 +23,7 @@ from typing import List
 
 # Make sure the version of sockeye being tested is first on the system path
 try:
-    import contrib.autopilot.autopilot as autopilot
+    import sockeye_contrib.autopilot.autopilot as autopilot
 except ImportError:
     SOCKEYE_ROOT = os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
     PYTHONPATH = "PYTHONPATH"
@@ -32,7 +32,7 @@ except ImportError:
     else:
         os.environ[PYTHONPATH] = SOCKEYE_ROOT
     sys.path.append(SOCKEYE_ROOT)
-    import contrib.autopilot.autopilot as autopilot
+    import sockeye_contrib.autopilot.autopilot as autopilot
 
 
 # Test-specific constants
@@ -84,7 +84,7 @@ def main():
         # WMT task with raw data (Transformer)
         command = [sys.executable,
                    "-m",
-                   "contrib.autopilot.autopilot",
+                   "sockeye_contrib.autopilot.autopilot",
                    "--task={}".format(WMT_TASK),
                    "--model=transformer",
                    "--gpus=0",
@@ -94,7 +94,7 @@ def main():
         # WMT task with raw data (GNMT)
         command = [sys.executable,
                    "-m",
-                   "contrib.autopilot.autopilot",
+                   "sockeye_contrib.autopilot.autopilot",
                    "--task={}".format(WMT_TASK),
                    "--model=gnmt_like",
                    "--decode-settings=gnmt_like",
@@ -107,7 +107,7 @@ def main():
         # WNMT task with pre-tokenized data (Transformer)
         # command = [sys.executable,
         #            "-m",
-        #            "contrib.autopilot.autopilot",
+        #            "sockeye_contrib.autopilot.autopilot",
         #            "--task={}".format(WNMT_TASK),
         #            "--model=transformer",
         #            "--gpus=0",
@@ -117,7 +117,7 @@ def main():
         # WMT task, prepare data only
         command = [sys.executable,
                    "-m",
-                   "contrib.autopilot.autopilot",
+                   "sockeye_contrib.autopilot.autopilot",
                    "--task={}".format(DATA_ONLY_TASK),
                    "--model=none",
                    "--gpus=0",
@@ -127,7 +127,7 @@ def main():
         # Custom task (raw data, Transformer)
         command = [sys.executable,
                    "-m",
-                   "contrib.autopilot.autopilot",
+                   "sockeye_contrib.autopilot.autopilot",
                    "--custom-task=custom_raw",
                    "--custom-train",
                    os.path.join(work_dir, autopilot.DIR_SYSTEMS, WMT_TASK + autopilot.SUFFIX_TEST, autopilot.DIR_DATA,
@@ -156,7 +156,7 @@ def main():
         # Custom task (tokenized data, Transformer)
         command = [sys.executable,
                    "-m",
-                   "contrib.autopilot.autopilot",
+                   "sockeye_contrib.autopilot.autopilot",
                    "--custom-task=custom_tok",
                    "--custom-train",
                    os.path.join(work_dir, autopilot.DIR_SYSTEMS, WMT_TASK + autopilot.SUFFIX_TEST, autopilot.DIR_DATA,
@@ -183,7 +183,7 @@ def main():
         # Custom task (byte-pair encoded data, Transformer)
         command = [sys.executable,
                    "-m",
-                   "contrib.autopilot.autopilot",
+                   "sockeye_contrib.autopilot.autopilot",
                    "--custom-task=custom_bpe",
                    "--custom-train",
                    os.path.join(work_dir, autopilot.DIR_SYSTEMS, WMT_TASK + autopilot.SUFFIX_TEST, autopilot.DIR_DATA,
