+++ sockeye_contrib/fast_align/lex_table.sh
