+++ sockeye_contrib/vistools/README.md
@@ -7,7 +7,7 @@ This package generates javascript-based graphs of the beam search performed by s
 First, install the dependencies required for the visualizations:
 
 ```sh
-pip install -r contrib/vistools/requirements.txt
+pip install -r sockeye_contrib/vistools/requirements.txt
 ```
 
 ### Store the beam histories
@@ -27,7 +27,7 @@ python3 -m sockeye.translate --models model \
 After inference, the graphs can be generated with:
 
 ```
-python3 contrib/vistools/generate_graphs.py -d beams.json -o generated_graphs
+python3 sockeye_contrib/vistools/generate_graphs.py -d beams.json -o generated_graphs
 ```
 
 The `generated_graphs/` folder will contain one `html` file per sentence. Opening it in your browser will show the interactive graph.
