+++ sockeye_contrib/vistools/templates/tree.js
