+++ sockeye_contrib/vistools/test/resources/beams.json
