+++ test/unit/test_bleu.py
@@ -15,7 +15,7 @@ from collections import namedtuple
 
 import pytest
 
-from contrib import sacrebleu
+from sockeye_contrib import sacrebleu
 
 EPSILON = 1e-8
